==> 0_mibrx.ipynb <==
import pandas as pd
from urllib.request import urlopen
import warnings
import csv
warnings.filterwarnings("ignore")

ibrxdia = pd.read_csv("IBXXDia_29-07-22.csv", encoding="cp1252", sep = ';',skiprows=1)
ibrxdia = ibrxdia.index
ibrxdia = list(ibrxdia)
#Filtra apenas tickers
ibrxdia = [w for w in ibrxdia if w.isupper()]
ibrxdia
#Renomear
ibrxmes = ibrxdia
ibrxmes

len(ibrxmes)

ibrx =pd.read_excel("ibrx.xlsx", sheet_name='Sheet1', index_col='Data')
#ibrx.T.index.unique()

todosativos = ibrx.T.index.unique()
#todosativos

#verifica se há empresa nova no IBRX?
print(list(set(ibrx) - set(todosativos)) == [])
print(len(list(set(todosativos) - set(ibrx))) + len(set(ibrx)) == len(list(set(todosativos))))
#True = Não!
#False = sim!

#CRIA UMA COLUNA NOVA PARA A EMPRESA SE ELA NÃO EXISTE
novaempresa = list(set(ibrxmes) - set(todosativos))
ibrxnova = ibrx
for i in novaempresa:
    ibrxnova[i] = 0
    
#ibrxnova[novaempresa]

#LENDO O DIA NA PRIMEIRA LINHA DO ARQUIVO

with open("IBXXDia_29-07-22.csv", encoding="cp1252") as csv_file:
    data=csv.reader(csv_file)
    for row in list(data)[0:1]:
        dia = row
        
dia = str(dia)
dia = dia[25:33]
dia = dia.replace('22','2022')

hoje = dia
dia

#INSERE 1 NAS EMPRESAS DO IBRX NOVO
z = ibrxnova.T


z[hoje] = 0
z
for i in ibrxmes:
    z[hoje].T[i] = 1


z = z.T
z

z.to_excel('ibrx.xlsx')

z
