==> 3_financeiro.ipynb <==
# Dados Financeiros

## Extraindo e Carregando os Dados

#Importando pacotes
import pandas as pd
import numpy as np
from urllib.request import urlopen
import missingno as msno
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
from scipy.stats import mstats
warnings.filterwarnings("ignore")

#Descriminando URL de download
url = "https://api.data.economatica.com/1/oficial/datafeed/download/1/Y5gu39g8DE8AXXQX0PSBo%2BwtJRdw0T4fZ7zIh%2FHGKlIUZcmuZn1CcyK5DhsRLXLWNjA5pxiND4Bv979MBmiP1v23jt6Ar684VGgmPrrisdumh0lTrtS%2BrrEVG1kAD4AGaRiAQe7eHCujrDcLtxYd%2BqRGFc1h6Wf1nrbG7DSYeSbnGpu8UOsb0jW%2BBrFkKP9IXAar0cXk0Ay0WxvthVP%2FY2m7qPyeoMPocIQnXvZl94crwrBy%2FpOx8IRWw9DY7C36aUGeBwL5AQOFCT%2FVjr4fX0479MLx4Ct4GFCxPi2XkPdpWlcODCeRmytAS9kth5jCrY1nkUlskeg%2FfisJvWqDwA%3D%3D"

%%time
#Abrindo a base
data = pd.read_csv(urlopen(url),encoding='latin-1')
#Retirando sufixo
data['Ativo']=data['Ativo'].str.rstrip('<XBSP>')
#Declarando coluna data como datatime
#data['data'] = pd.to_datetime(data['data'])
#Renomeando colunas
listacolunas=['ativo','data',\
             "lpa12m",\
"lpa3m",
"lpaex",
"vpa",
"vdacao12m",
"vdacao3m",
"vdacaoex",
"ebitdaacao12m",
"ebitdaacao3m",
"ebitdaacaoex",
"divttlq",
"divttbr",
"divbrativo",
"divbrpatr",
"dlpl",
"estcap",
"ebitdivbr12m",
"ebitdivbr3m",
"ebitdivbrex",
"ebitdivlq12m",
"ebitdivlq3m",
"ebitdivlqex",
"ebitdespfbr12m",
"ebitdespfbr3m",
"ebitdespfbrex",
"ebitdespflq12m",
"ebitdespflq3m",
"ebitdespflqex",
"divbrebitda12m",
"divbrebitda3m",
"divbrebitdaex",
"divlqebitda12m",
"divlqebita12m",
"divlqebitdaex",
"divcpdivtt",
"exgttat",
"exgttpl",
"exgttrec12m",
"exgttrec3m",
"exgttrecex",
"afxpl",
"invpl",
"depvdept",
"repaspsv",
"liqger",
"liqcor",
"liqsec",
"capgir",
"capemp",
"pmestq12m",
"pmestq3m",
"pmestqex",
"pmforn12m",
"pmforn3m",
"pmfornex",
"pmrecb12m",
"pmrecb3m",
"pmrecbex",
"cicfin12m",
"cicfin3m",
"cicfinex",
"cicope12m",
"cicope3m",
"cicopeex",
"ebitda12m",
"ebitda3m",
"ebitdaex",
"lairdfl12m",
"lairdfl3m",
"lairdflex",
"giroat12m",
"giroat3m",
"giroatex",
"giropl12m",
"giropl3m",
"giroplex",
"mgbr12m",
"mgbr3m",
"mgbrex",
"mgebit12m",
"mgebit3m",
"mgebitex",
"mglq12m",
"mglq3m",
"mglqex",
"mgebitda12m",
"mgebitda3m",
"mgebitdaex",
"roa12m",
"roa3m",
"roaex",
"roefim12m",
"roefim3m",
"roefimex",
"roemed12m",
"roemed3m",
"roemedex",
"roein12m",
"roein3m",
"roeinex",
"roicfim12m",
"roicfim3m",
"roicfimex",
"roicmed12m",
"roicmed3m",
"roicmedex",
"roicin12m",
"roicin3m",
"roicinex",
"capinv",
"reinsu12m",
"reinsu3m",
"reinsuex",
"girodep12m",
"girodep3m",
"girodepex",
"dvrof12m",
"dvrof3m",
"dvrofex",
"alavanfin12m",
"alavanfin3m",
"alavanfinex",
"alavanope12m",
"alavanope3m",
"alavanopeex",
"capex12m",
"capex3m",
"capexex",
"depramor12m",
"depramor3m",
"depramorex",
"capexdepr12m",
"capexdepr3m",
"capexdeprex",
"deprimob12m",
"deprimob3m",
"deprimobex",
"fcl12m",
"fcl3m",
"fclex"
]
data.columns = listacolunas

#Confirmando o número de empresas
unique=data['ativo'].unique()
len(unique)

#data

new = data["data"].str.split("-", n = 1, expand = True)
data["mes"]= new[0]
data["ano"]= new[1]
data["mes"]=data["mes"].replace({"Jan":1,"Fev":2,"Mar":3,"Abr":4,"Mai":5,"Jun":6,"Jul":7,"Ago":8,"Set":9,"Out":10,"Nov":11,"Dez":12})
data['date'] = pd.to_datetime(data['ano'].astype(str) + data['mes'].astype(str).str.zfill(2), format='%Y%m')
data['date'] =data['date'] + pd.offsets.MonthEnd()
listadata=pd.date_range(start=data['date'].min(), end=pd.Timestamp.today()  , freq=pd.offsets.MonthEnd(1))
data["data"]=data["date"]
data.drop(["date", "ano", "mes"], axis=1, inplace=True)
#listadata

listadata

data=data.sort_values(['data', "ativo"], ascending=[True, True])


#Declarando hifens como dados faltantes NAN
data[data == "-"] = np.nan
#Preenchendo valores ausentes (médoto para frente (ffill))
data = data.set_index('ativo', append=True)
data = data.groupby(level=1).ffill()
data = data.reset_index(level=1)

#Preenchendo valores ausentes (médoto para trás (bfill))
#data = data.set_index('ativo', append=True)
#data = data.groupby(level=1).bfill()
#data = data.reset_index(level=1)

#Preparando DataFrame de referência onde todos ativos terão todas as datas de fim de mes
index=pd.DataFrame({'data':listadata}) #usando a lista definida acima
indexx=pd.concat([index]*len(unique), ignore_index=True)
listaativo=np.repeat(unique,len(index))
indexx['ativo']=listaativo
#indexx

#Unindo a base de data/ativo completo com a base de dados (aparecerão Nan's nas datas sem valor)
financeiro=pd.merge(indexx, data, how="left", on=["ativo", "data"])

#Preenchendo valores ausentes (médoto para frente (ffill))
financeiro = financeiro.set_index('ativo', append=True)
financeiro = financeiro.groupby(level=1).ffill()
financeiro = financeiro.reset_index(level=1)

#Salvando em excel
#financeiro.to_excel('financeito.xlsx')

#Compatibilizando datas 
data=pd.read_excel('data.xlsx')
data_index=data.loc[:,['data','ativo']]
#data_index

len(data_index)==len(financeiro)

list(data_index['ativo'])== list(financeiro['ativo'])

financeiro['data']=data_index['data']
#financeiro

#MATRIZ IBRX
ibrx=pd.read_excel('ibrx.xlsx', sheet_name='Sheet1', index_col='Data')
#ibrx

#Verificando igualdade de número de ativos
len(list(ibrx.columns))==len(unique)

#Verificando igualdade de NÚMERO de datas (o "-1" tira o mes de novembro de 2001 que tem incluído nos links mas não no IBRX)
len(list(ibrx.index))==len(index)-1

#Preparando DataFrame de referência do IBRX com data e ativos empilhados
empilhados=pd.Series(ibrx.values.ravel('F'))#Empilhando colunas do DataFrame
indexibrx=pd.DataFrame({'data':list(ibrx.index)}) #Criando um índice
indexxibrx=pd.concat([indexibrx]*len(list(ibrx.columns)), ignore_index=True)
listaativo=np.repeat(list(ibrx.columns),len(indexibrx))
indexxibrx['ativo']=listaativo
indexxibrx['IBRX']=empilhados

#Unindo base de dados com IBRX
financeiro1=pd.merge(indexxibrx, financeiro, how="left", on=["ativo", "data"])
#financeiro1

financeiro1.dtypes

#Formatando variáveis
financeiro1['ativo'] = financeiro1['ativo'].astype(str)
financeiro1.iloc[:,3:] = financeiro1.iloc[:,3:].astype(float)
financeiro1.dtypes

#Base de dados com participantes do IBRX 
#financeiro2=financeiro1[financeiro1['IBRX']==1]
financeiro2=financeiro1

financeiro2=financeiro2.sort_values(['data', "ativo"], ascending=[True, True])
#financeiro2

#financeiro2.to_excel("financeiro2.xlsx")

#Outra análise visual de missing
#financeiro2=financeiro2.sort_values(['data', "ativo"], ascending=[True, True])
#balanco2=balanco2.sort_values(['data'], ascending=[True])
#sns.heatmap(financeiro2.isnull(), cbar=False)
#sns.heatmap(price2.isnull(), cbar=False,yticklabels=list(price2['ativo']))

financeiro2=financeiro2[['ativo','data',\
                         "IBRX",\
              "lpa12m",\
              "vdacao12m",\
              "ebitdaacao12m",\
              "divttlq","divttbr","divbrativo","divbrpatr","dlpl",\
              "estcap",\
              "ebitdivbr12m",\
              "ebitdivlq12m",\
              "ebitdespfbr12m",\
              "ebitdespflq12m",\
              "divbrebitda12m",\
              "divlqebitda12m",\
              "divcpdivtt",\
              "exgttat","exgttpl",\
              "exgttrec12m",\
              "afxpl","invpl","liqger","liqcor","liqsec","capgir","capemp",\
              "pmestq12m",\
              "pmforn12m",\
              "pmrecb12m",\
              "cicfin12m",\
              "cicope12m",\
              "ebitda12m",\
              "lairdfl12m",\
              "giroat12m",\
              "giropl12m",\
              "mgbr12m",\
              "mgebit12m",\
              "mglq12m",\
              "mgebitda12m",\
              "roa12m",\
              "roefim12m",\
              "roemed12m",\
              "roein12m",\
              "roicfim12m",\
              "roicmed12m",\
              "roicin12m",\
              "capinv",\
              "alavanfin12m",\
              "alavanope12m",\
              "capex12m",\
              "depramor12m",\
              "capexdepr12m",\
              "deprimob12m",\
              #"fcl12m" \
                        ]]
financeiro2

#Outra análise visual de missing
#financeiro2=financeiro2.sort_values(['data', "ativo"], ascending=[True, True])
#balanco2=balanco2.sort_values(['data'], ascending=[True])
#sns.heatmap(financeiro2.isnull(), cbar=False)
#sns.heatmap(price2.isnull(), cbar=False,yticklabels=list(price2['ativo']))

#financeiro2.to_excel("financeiro2.xlsx")

## Análise Exploratória de Dados

### Missing data

#Análise detalhada da base de dados
#from pandasgui import show
#show(financeiro2)

#Métricas para análise explanatória inicial
def estats(df):
    a=list(df.count()) #nonnullvalues
    
    b=list(df.isnull().sum() * 100 / len(df)) #percent_missing
   
    c = []#firstnonnullline
    for i in df.columns:
        y = df[i].first_valid_index()
        c.append(y) 
   
    d = [] #date_firstnonnullline
    for i in c:
        try:
            y = df.loc[i,'data']
            d.append(y)
        except Exception:
            y = np.nan
            d.append(y)

    e = [] #uniquevalues
    for i in df.columns:
        y = df[i].nunique()
        e.append(y)

    #f= list(map(lambda x,y: (x/y)*100, e,a)) #percentunique

    g = [] #zerocount
    for i in df.columns:
        y = (df[i] == "0").sum()
        g.append(y)
    
    h= list(map(lambda x,y: (x/y)*100, g,a)) #percentzero   
    
    #Selecionando algumas métricas mais relevantes
    estats=pd.DataFrame(data={'percent_missing': b, 'date_firstnonnullline': d, 'percentzero':h },index=df.columns)
    
    return estats.iloc[3:,:] #mostrando as métricas somente para as variáveis de fato

#Mostrando as métricas para toda base de dados
estats(financeiro2)

#Olhando para um ticher em especial
#financeiro2[financeiro2['ativo']=='ABEV3']

#Olhando para uma data em especial
#financeiro2[financeiro2['data']=='2001-12-28']

#Mostrando as métricas para um ticker em especial
#estats(financeiro2[financeiro2['ativo']=='ABEV3'])

#Mostrando as métricas para todos os tickers
#todos= list(map(lambda x: estats(financeiro2[financeiro2['ativo']==x]), list(financeiro2['ativo'].unique())))
#estatsall=pd.concat(todos, axis=0, keys=list(financeiro2['ativo'].unique()))
#estatsall

#Análise visual de missing
#df=financeiro2.copy()
#df = df.sort_values(by=['data'])

#fontsize = 20
    
#fig, ax = plt.subplots(1, 1, figsize=[10, 10]) #mude aqui o tamanho da matriz
#msno.matrix(df=df.iloc[:,3:], ax=ax, color=(0.2, 0.2, 0.2), sparkline=False, fontsize=fontsize)

#label = list(df['ativo'].unique())
#ylim_start, ylim_end = ax.get_ylim()
#step_size = df.shape[0] / len(label)
#_ = ax.yaxis.set_ticks(np.arange(ylim_end, ylim_start, step_size))
#_ = ax.yaxis.set_ticklabels(label, fontsize=fontsize)

financeiro31=financeiro2[['ativo','data',\
                         "IBRX"]]
#financeiro31

financeiro32=financeiro2[['ativo','data',\
                         "lpa12m",\
              "vdacao12m",\
              "ebitdaacao12m",\
              "divttlq","divttbr","divbrativo","divbrpatr","dlpl",\
              "estcap",\
              "ebitdivbr12m",\
              "ebitdivlq12m",\
              "ebitdespfbr12m",\
              "ebitdespflq12m",\
              "divbrebitda12m",\
              "divlqebitda12m",\
              "divcpdivtt",\
              "exgttat","exgttpl",\
              "exgttrec12m",\
              "afxpl","invpl","liqger","liqcor","liqsec","capgir","capemp",\
              "pmestq12m",\
              "pmforn12m",\
              "pmrecb12m",\
              "cicfin12m",\
              "cicope12m",\
              "ebitda12m",\
              "lairdfl12m",\
              "giroat12m",\
              "giropl12m",\
              "mgbr12m",\
              "mgebit12m",\
              "mglq12m",\
              "mgebitda12m",\
              "roa12m",\
              "roefim12m",\
              "roemed12m",\
              "roein12m",\
              "roicfim12m",\
              "roicmed12m",\
              "roicin12m",\
              "capinv",\
              "alavanfin12m",\
              "alavanope12m",\
              "capex12m",\
              "depramor12m",\
              "capexdepr12m",\
              "deprimob12m",\
              #"fcl12m" \
                        ]]
#financeiro32

financeiro4=financeiro32 ## Rodar apenas uma vez
financeiro4 = financeiro4.set_index('data', append=False) # data como índice para possibilitar shift
financeiro4 = financeiro4.set_index('ativo', append=True)
#balanco4 = balanco4.groupby(level=1).ffill()
financeiro4 = financeiro4.groupby(level=1).shift(3) # Shift para defasagem de 3 meses para dados balanço
financeiro4 = financeiro4.reset_index(level=0) # Reset índice - 2 vezes level 0 para manter ordem das colunas
financeiro4 = financeiro4.reset_index(level=0)
financeiro4.head()

financeiro4=financeiro31.merge(financeiro4, on=["data", "ativo",])
financeiro4

financeiro2[financeiro2["ativo"]=="ABEV3"].tail()

financeiro4[financeiro4["ativo"]=="ABEV3"].head()

#Análise visual de missing
#df=financeiro4.copy()
#df = df.sort_values(by=['data'])
#fontsize = 12    
#fig, ax = plt.subplots(1, 1, figsize=[15, 5.5]) #mude aqui o tamanho da matriz
#msno.matrix(df=df.iloc[:,3:], ax=ax, color=(0.2, 0.2, 0.2), sparkline=False, fontsize=fontsize)

#Salvando em excel
financeiro4.to_excel('fin.xlsx')

financeiro4
