==> 5_tecnico.ipynb <==
# Dados Técnicos

## Extraindo e Carregando os Dados

#Importando pacotes
import pandas as pd
import numpy as np
from urllib.request import urlopen
import missingno as msno
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
from scipy.stats import mstats
warnings.filterwarnings("ignore")

#Descriminando URL de download

#tecnicos basicos
url1 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/Y8aC8bcrAHj9aycRMj3XkKiiOBmdyWOGZa0f8oDr9oFfIHxLFNjwAIRJl06M45oitpSeenVejZhr85e7ATu2LqhM9sDF7ORzfs%2BIJjLX9gy7W2ScGh3lhXv0Axk33FCgBQtE0Zn4q5%2F%2BTgd7fu9zvGKOQsZr6GHUo17uhsu%2BIuY%2FWukm6KUcUg0LyuVH0ZXRAUENUMZJNxU0KfSVFEL%2Fstnn9B7uX%2FZ%2B7Pyy3YJTONadwhoWNixlLZJHycGtxkMHkHKxJwd7%2BuKykOvBqPQW7L3PLQtQ1eBlMA3QWwkB4VAxGZ0AOcyuD9xhzhsLH4A9GSy8uzYq7hnlDGodur%2FkcA%3D%3D"
#tecniscos liquidez
url2 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/kcMxNpAbRQfB3mklSFm4PrE9AdYeRvxi5yNdiebBy4G2QD8bZPKWIA4QqNKjOdkjGg5Hh5rDRjl4nPp%2FqZeg8kqIaMkyy2CaCFmIr2sFGahm3Dfrs%2FA9XWNxY%2BwnBptKJvLx0on3Hh2KsYmhVHS35AJx5T%2Fd8mEz32p%2F3s%2BoT%2Bf3YUojRMkuhZJZUL61z32QADJM%2Fej7l7eJxXJtVKvjuy4YpLXcY%2FKzzN4H%2BGTW8GZd23ugxRK2IOk%2FXZQLVx1YEAN28pp0zCIit7zn0Z1RTOM8%2FG00aTZg%2BP3hJphg8AEMUELisRCWGnzWeNXJuTIKiS%2FPFevVY2PwtlVyIVqijg%3D%3D"
#tecnico1
url3 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/EPRnl%2BOATda%2Bl22IUH77DSdr49Ng%2B9SmPv03eXsd1bS3Wz55cdFpy0v8AlKEyQB4QhIVV9AmfGfb840DFV26ERMs0P%2F%2BPxPqf0kAs3di4TAn7sWrO3J56nd78lSzGYq%2Bzzd16LmuuxuMXRcuSBx0BXIe7Hi9%2BKcBjyaXREixKa%2F86Qk4vTEXxm2K%2FgSUQi6vNhCTk2mkXbMErRL5gBnlkhEVf53Rfhe3HrLkyah2RzGeiJlokGc8amdgdVRJqRh71dtxkE6ZtBFoguphEeZDBmDZlMHrf%2FtG2uu4AAb9WxZpM2klLdwnrkC%2Fh7EG5QzuYsdxU9flN4IHtdJemHVQWg%3D%3D"
#tecnico2
url4 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/OsqHvQggqV931x5YvHiD0KiAPpEe606H6i%2FandrFNMgHxGs446GHin7XQetbCjCUYnCUeCEZf116hEFNo3v6gjewm5UcUxeBvvYBvd10RWlBjAwf5%2BEUrnEScNRH%2F7tV3prUhQEOnWMUXWcujtfsQ5QZEZy0V5Bq%2BasNpy%2FEWb9jNOryUQjDKv8pLRnP1sQkewjD046miRarQ1lU5e7MDNg6Q30sakWtiISJAqUdKrzSVFdk5IIBQEEhavVj10tCZtlYOqP4xJhl7dlWLXkpg%2FAtUE46vsX1ti1mRmxpQJncn%2FUcuwBAi6BLjY1u0tAk1cEr835SXsBOXpqv%2FIOXEg%3D%3D"
#tecnico4
url5 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/iL5guqIjJX9g56BDv2vuRdpy4C89fD7v%2BVoo63UOItPD%2F7IHs8ItHz7p5JzcNOywJTTVnJFLpdx3OlDK9VzcvGVqwLsRTzL9LBMAqm7QYVtuZoGv9pcU2dPMdUeROsmQv%2FtkfLb6DtsPcNHJIZr4uhClWEOaTJLDdbn1IqYXSca7yFsBVcacPcyDn04w%2BO2s2dv%2B%2BsaWta0tA50ZJ10COzQt4iqAomNyJVAKFdjxHuIG9xh4iAQwQqlikL36TsGNZgNEQeCrL2TNqAjxn4f22hVsf1PWiCjrch%2BthXmF4KenmH3VrOD%2FqzofXWbRN3S6%2FjmJL8WV7MxEfWvgDlaUHA%3D%3D"
url6 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/ttdebv9FXLXhnhHdoi1VMC%2BVHTxjt4dVl%2BcfRjBlXunjE2Ueo7%2BQ9UJJEFrzsgx1MIQNvZ6I45EGYtTfKFFWsWIWa9EwZkWYGLJxsUPrKGDtcYnCBrDv%2BUEw%2F0ah91gHI96XeIE83pnyxJnsY6JZbt5nEab7B3B8r8YXbdeBjy3%2FNbr2vaKoMzQ7XgMWP9q8SQL%2F3UVl39LdiK9UOBMqUmnH47djdR0Uc2lWRcO2dp2rnYBfvKqQteiMHR%2FNf7jkyUDOQG4MiGe%2FcGZcnvNch50PorVqvgkcPOcA7zQ7NFCigJYrZ9XaJjvI2%2BZ1Erc49gSyVS8utI6sbFUG%2FJgETA%3D%3D"
url7 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/BU71zNb3EuMhxq2iLsD42%2BBT1e4renG6x8cqOSUwPLmj45LQ4VPFOGCZ0Oua%2BGc%2Ff2sL4MTKw8K1VJcPKJWjZhdSIti0EIY5fib21Uqb9%2FbAzF%2BVlqGGf%2Fip8YKqGbsC9m6zdqElxSi%2FjwLqDN8UjYONp16vDeAHjTRWgViuMUalQJcDJOAISIQnhv1q6500AWs8SqStWPtJsSPoWjldNUeL2%2Fx41XewIPjvtmH9JnnwpU6PttRUbQ094MWSlqqHbl7SUxBhh3s%2BY6lvROx4KG%2BVejQGpjId2v9KArw%2B%2BZ4AdhobQoivpFIKOyusj%2BmUgo1aMsdFeX9QTp%2B1oALpww%3D%3D"
url8 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/DNIVQNdWIyScCNlr%2BrSEnDsgDK91W1rG7PLi1UmiUybuM2%2Bx8SQydlSrBrd%2F7zWxUWkOzEUu%2FDmFzlz54ZCCds%2BMrrrxI8qW8xofb4Fx5aKM0DYo6BcAhFDnW4fNZS3kSG7JMkC9CnBvwJqlsZsetMlY74KD6%2B2AqkHaQylm3mpY1Thd3f28B7cLLDzIt%2B84bXhCm4vNGW9vKhsBy4yY%2BHVTrgvEY2aiGYqr0N5sdKeK6O6%2B8h9oTXXso7J2Gcdyq96dgwlTYKzjGt9w%2FfWrZZ5z8PRRyE%2B8yDl22EfbsYjbcONl6aRsVj5JugPCCNF6vC55w5XquhcnG7B5pRfAHQ%3D%3D"
url9 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/jC0NfwGuWPt1HgTs90U1dt%2BSloVQ8jclurkd1qeGDpnLIr2cjOJpRKxH7rTU0Yhn16w5VKsk1fGwQSZFQfVnugLsZxrheTzdYhrNtczBL%2FT21WAY8vnSE45hdib2mzC%2FC%2BY1dAr7EcZKhEDBcvpKdxrfvAUVRXJOEopoPstiq4Qe%2BT8bmtBT69rWcCuH3a4oCvxnUN1NcLRc0RIGEAxZ9XWRlOz21Bh%2FHz9dnSnvrAN%2F%2BzJqvbVms8T6sg%2FNh4AmhYTR6mJG1x%2B4m1pm4b9e7gqTpFsK3wDTxgoU8EauV1JyC9oocVYlXot7BiXgvH7tuYqlRXGo2HPIIn0gr%2BZQCw%3D%3D"
url10 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/XDb1V7PZzgWWCKWB%2BkwIfbQ6A9z%2FhLJ7nrcxQFlBhzE%2FKT2sNNjNof%2BVJ9oB4l0uHwzdun8s%2FubxgCVSayLmfyR99WlrUG1Xj8qJmG4tzD9Bta6MbpgHfYdIyBNBQSm2BjI77KSbr77fBRBgqGQHHxB83t%2BwT0Qgh6VLzYgbSL6IQmkl2BgavrIb5S9Hxqr8QFZ%2BCW54tXLhvrwv2hiOnvQ4eXXcL5sR8uWv8H4JYDB9KubzJsBQGGK1Kz7YKrzjMEaSHapGv0Lh%2F1Otk%2BaH9IS%2FF3Zpz1lWiA5eK0RK9yHRgtRZafQKmYiTlSwJhc43I0LJtxTfAZv5%2Fhiwuqd9GA%3D%3D"
url11 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/RB%2B0%2FOZyUkIAoMz%2F6pyR6rP6bUHq1B1PmDMqfoXTdz6W5%2BQOVZ0ZN%2B4xFj%2B3OAkIISXDhaJwvHk4jH4fB5ow%2Flu9kkFDKW%2F%2BgkIc6YDw4IrRqubY9V10WvHTEmA1txadfHl649VSBJpYv1IgDSJ0mJ0S0HFhuKa7UxmWxvWtWpuWlrf%2B2HiamblYgU89G5PFgCDY96FGKaFsQLW3Abd4X7qeLYZ3ucj6i8aHpgqIoZOEapOyxczAFsaXtNEalpymwvGQKi7xSgrQZlk4DPq3VW%2F5Iy7xLr63Cy6eIsZQJgDtTaZVyJ9jLwGtztalq1EfwtZmjuBL1KyHePU5k6DuzQ%3D%3D"
#tecnico3
url12 = "https://api.data.economatica.com/1/oficial/datafeed/download/1/K%2Fp2eQy2aUX77W4Qml7%2FhyyacIkKebXRuSV49svE25knX4TAg6NYAJpzX81yk70Cm6iPLCb8qFel0HJAIzT1wgJY%2BJ9OMsR6mchkzjbW59SbFcV8WtpbHv%2B6DmhFzS1li7Fzo3idetVB3ytaLBl4GYl12DGYfHQVwqe%2BbE1GhK9ALnbxcufcWG754iGX8E47lbOTGuZGyp%2BlypiwCR%2BTefmWTEyJ%2Bwk0ZGLBVwy%2FWWbJ0IavJgc%2BmaQ%2ButocGH9iQoS0IO3Q1EO4g2D%2Bg%2B2kGPwMGb4kscRqLeOebmA9JydFPnJd7%2FPRjOSLdtXgO8NlMoU5kcMyOUcWg8kHAg%2BnpQ%3D%3D"

indices = "https://api.data.economatica.com/1/oficial/datafeed/download/1/jIFbIzZ6mj3itlfxKk9y5RW6yi7UT2OuKsIfrip3AbmQn3hIqrNgLKl0NmC3dkb5S4JISEgAD2dNFmrBs7bH%2BSOq4iCd%2FRwDRm9p5x7rtnWyUFAzJJJBQmUMeM48%2BzvYDrkHNHFEkvWGM4ifq%2BJ48jfgvk7dn%2BVGUKkcZxI4FjA4yXM%2FoFjWVpwgeOLBLs%2FpCuttUBMntHZfSWm%2BQv2BR9t%2FwlQPJyvEyvzldFtsC6ugE%2B4EOxGywjw1%2B037t2j0UxZTHoJr1Z4f%2FCXO%2BXGUi1FP7xepIiAmnQLxs3bS8uR9lcTWeRQh9P6EbuqfEY8OC1lz%2B0kAgwoRe65rWMZYHQ%3D%3D"

%%time
#TÉCNICOS
#Abrindo a base
data1 = pd.read_csv(urlopen(url1),encoding='latin-1')
data2 = pd.read_csv(urlopen(url2),encoding='latin-1')
data3 = pd.read_csv(urlopen(url3),encoding='latin-1')
data4 = pd.read_csv(urlopen(url4),encoding='latin-1')
data5 = pd.read_csv(urlopen(url5),encoding='latin-1')
data6 = pd.read_csv(urlopen(url6),encoding='latin-1')
data7 = pd.read_csv(urlopen(url7),encoding='latin-1')
data8 = pd.read_csv(urlopen(url8),encoding='latin-1')
data9 = pd.read_csv(urlopen(url9),encoding='latin-1')
data10 = pd.read_csv(urlopen(url10),encoding='latin-1')
data11 = pd.read_csv(urlopen(url11),encoding='latin-1')
data12 = pd.read_csv(urlopen(url12),encoding='latin-1')
#Retirando sufixo
data1['Ativo']=data1['Ativo'].str.rstrip('<XBSP>')
data2['Ativo']=data2['Ativo'].str.rstrip('<XBSP>')
data3['Ativo']=data3['Ativo'].str.rstrip('<XBSP>')
data4['Ativo']=data4['Ativo'].str.rstrip('<XBSP>')
data5['Ativo']=data5['Ativo'].str.rstrip('<XBSP>')
data6['Ativo']=data6['Ativo'].str.rstrip('<XBSP>')
data7['Ativo']=data7['Ativo'].str.rstrip('<XBSP>')
data8['Ativo']=data8['Ativo'].str.rstrip('<XBSP>')
data9['Ativo']=data9['Ativo'].str.rstrip('<XBSP>')
data10['Ativo']=data10['Ativo'].str.rstrip('<XBSP>')
data11['Ativo']=data11['Ativo'].str.rstrip('<XBSP>')
data12['Ativo']=data12['Ativo'].str.rstrip('<XBSP>')

#Verificando dimensões
print(data1.shape,data2.shape,data3.shape,data4.shape,data5.shape, data6.shape,\
      data7.shape,data8.shape,data9.shape,data10.shape,data11.shape,data12.shape)

#Declarando coluna data como datatime
data1['Data'] = pd.to_datetime(data1['Data'])
data2['Data'] = pd.to_datetime(data2['Data'])
data3['Data'] = pd.to_datetime(data3['Data'])
data4['Data'] = pd.to_datetime(data4['Data'])
data5['Data'] = pd.to_datetime(data5['Data'])
data6['Data'] = pd.to_datetime(data6['Data'])
data7['Data'] = pd.to_datetime(data6['Data'])
data8['Data'] = pd.to_datetime(data6['Data'])
data9['Data'] = pd.to_datetime(data6['Data'])
data10['Data'] = pd.to_datetime(data6['Data'])
data11['Data'] = pd.to_datetime(data6['Data'])
data12['Data'] = pd.to_datetime(data6['Data'])

#Reunindo os dados
data=pd.concat([data1,data2.iloc[:,2:],data3.iloc[:,2:],data4.iloc[:,2:],data5.iloc[:,2:],data6.iloc[:,2:],\
               data7.iloc[:,2:],data8.iloc[:,2:],data9.iloc[:,2:],data10.iloc[:,2:],data11.iloc[:,2:],data12.iloc[:,2:]], axis=1)
#Renomeando variáveis
listacolunas = ["ativo","data","ret1m","ret3m","ret6m","ret12m","divdacao1a","dyieldfim1a","dyieldinic1a",\
                "dyieldult12m","liq1m","liq3m","liq6m","liq12m","negocliq1m","negocliq3m","negocliq6m","negocliq12m","presbolsa1m","presbolsa3m",\
                "presbolsa6m","presbolsa12m","vmedio1m","vmedio3m","vmedio6m","vmedio12m","vol1m","vol3m","vol6m","vol12m","desvp1m",\
                "desvp3m","desvp6m","desvp12m","track1m","track3m","track6m","track12m","vatrisk1a","vatrisk2a","vatrisk3a","alfa1a",\
                "alfa2a","alfa3a","beta6m","beta1a","beta2a","beta3a","beta5a","corr1m","corr3m","corr6m","corr1a","corr2a",\
                "corr3a","sharp1m","sharp3m","sharp6m","sharp1a","sharp2a","sharp3a","sortin1m","sortin3m","sortin6m",\
                "sortin1a","sortin2a","sortin3a","inforatio1m","inforatio3m","inforatio6m","inforatio1a","inforatio2a",\
                "inforatio3a","treyn1m","treyn3m","treyn6m","treyn1a","treyn2a","treyn3a","maxserie1m","maxserie3m","maxserie6m","maxserie1a",\
                "minserie1m","minserie3m","minserie6m","minserie1a","maxdrdcrec1m","maxdrdcrec3m","maxdrdcrec6m","maxdrdcrec1a","maxdrdsrec1m",\
                "maxdrdsrec3m","maxdrdsrec6m","maxdrdsrec1a","retmax1a","retmin1a","retmed1a","qretpos1m","qretpos3m",\
                "qretpos6m","qretpos1a"]
data.columns=listacolunas

#Confirmando o número de empresas
unique=data['ativo'].unique()
len(unique)

data=data.sort_values(['data', "ativo"], ascending=[True, True])
#data.head(5)

#Declarando hifens como dados faltantes NAN
data[data == "-"] = np.nan
#Preenchendo valores ausentes (médoto para frente (ffill))
data = data.set_index('ativo', append=True)
data = data.groupby(level=1).ffill()
data = data.reset_index(level=1)

#Preenchendo valores ausentes (médoto para trás (bfill))
#data = data.set_index('Ativo', append=True)
#data = data.groupby(level=1).bfill()
#data = data.reset_index(level=1)

#IBOV que  ira auxiliar em construir um índice com datas do último dia do mês para o calendário brasileiro
#Abrindo a base
indice = pd.read_csv(urlopen(indices),encoding='latin-1')
#Retirando sufixo
indice['Ativo']=indice['Ativo'].str.rstrip('<XBSP>')
#Base somente com o IBOV
ibov=indice[indice['Ativo']=='IBOV'].iloc[:,:3]
#Declarando coluna data como datatime
ibov['Data'] = pd.to_datetime(ibov['Data'])
#Declarando hifens como dados faltantes NAN
ibov[ibov=='-']=np.nan
#Excluindo NAN (dias que a B3 não abriu)
ibov=ibov.dropna()
#Renomeia colunas
ibov.columns = ['ativo','data','fech']
#Identificando o último dia de cada mês
groups = ibov.groupby([ibov['data'].dt.year, ibov['data'].dt.month])['data']
last = []
for idx, group in groups:
    y = last.append(group.iloc[-1])
#last

#Marcando a última data disponível no objeto index e comparando com o ultimo dia do mês
import datetime
import calendar
from pandas.tseries.offsets import BMonthEnd
from pandas.tseries.offsets import BDay

today = datetime.date.today()
offset = BMonthEnd()
lastdaymt = offset.rollforward(today)
#lastdaymt_ifdec = offset.rollforward(today) - BDay(1)

#VERIFICA SE O ULTIMO DIA DO MÊS SERÁ 31 PARA O CASO DE DEZEMBRO, E APLICA A FORMULA CORRETA (TODOS OS MESES SERÃO A MESMA, EXCETO DEZEMBRO)
if (lastdaymt.day == 31) & (lastdaymt.month ==12):
    lastdaymt = offset.rollforward(today) - BDay(1)
else:
    lastdaymt = offset.rollforward(today)


#Preparando dataframe de referência onde todos ativos terão todas as datas de fim de mês
index=pd.DataFrame({'data':last}) #usando a lista definida acima#



if (index.iloc[-1][0] != lastdaymt):
    index=index[:-1] #tirando o mês que ainda não terminou
else:
    index=index
    
indexx=pd.concat([index]*len(unique), ignore_index=True)
listaativo=np.repeat(unique,len(index))
indexx['ativo']=listaativo
indexx

#Unindo a base de data/ativo completo com a base de dados (aparecerão Nan's nas datas sem valor)
tecnico=pd.merge(indexx, data, how="left", on=["ativo", "data"])
tecnico

#Salvando em excel
#tecnico.to_excel('tecnico.xlsx')

#MATRIZ IBRX
ibrx=pd.read_excel('ibrx.xlsx', sheet_name='Sheet1', index_col='Data')
ibrx

#Verificando igualdade de número de ativos
len(list(ibrx.columns))==len(unique)

#Verificando igualdade de NÚMERO de datas (o "-1" tira o mês de novembro de 2001 que tem incluído nos links mas não no IBRX)
len(list(ibrx.index))==len(index)-1

#Preparando dataframe de referência do IBRX com data e ativos empilhados
empilhados=pd.Series(ibrx.values.ravel('F'))#Empilhando colunas do dataframe
indexibrx=pd.DataFrame({'data':list(ibrx.index)}) #Criando um índice
indexxibrx=pd.concat([indexibrx]*len(list(ibrx.columns)), ignore_index=True)
listaativo=np.repeat(list(ibrx.columns),len(indexibrx))
indexxibrx['ativo']=listaativo
indexxibrx['IBRX']=empilhados

#Verificando igualdade entre datas do IBRX e da base DATA
list(indexibrx['data'])==list(index.loc[1:,'data'])

#Unindo base de dados com IBRX
tecnico1=pd.merge(indexxibrx, tecnico, how="left", on=["ativo", "data"])
tecnico1

tecnico1.dtypes

#Formatando variáveis
tecnico1['ativo'] = tecnico1['ativo'].astype(str)
tecnico1.iloc[:,3:] = tecnico1.iloc[:,3:].astype(float)
tecnico1.dtypes

#Base de dados com participantes do IBRX 
#tecnico2=tecnico1[tecnico1['IBRX']==1]
tecnico2=tecnico1

tecnico2=tecnico2.sort_values(['data', "ativo"], ascending=[True, True])
tecnico2.tail()


tecnico2=tecnico2[["ativo","data", "IBRX","ret1m","ret3m","ret6m","ret12m","divdacao1a","dyieldfim1a","dyieldinic1a",\
                "dyieldult12m","negocliq1m","negocliq3m","negocliq6m","negocliq12m",\
                "vmedio1m","vmedio3m","vmedio6m","vmedio12m","vol1m","vol3m","vol6m","vol12m",\
                "beta6m","beta1a","beta2a","beta3a","beta5a","corr1m","corr3m","corr6m","corr1a","corr2a",\
                "corr3a","sharp1m","sharp3m","sharp6m","sharp1a","sharp2a","sharp3a"]]
tecnico2

#tecnico2.to_excel("tecnico2.xlsx")

## Análise Exploratória de Dados

### Missing data

#Análise detalhada da base de dados
#from pandasgui import show
#show(tecnico2)

#Métricas para análise explanatória inicial
def estats(df):
    a=list(df.count()) #nonnullvalues
    
    b=list(df.isnull().sum() * 100 / len(df)) #percent_missing
   
    c = []#firstnonnullline
    for i in df.columns:
        y = df[i].first_valid_index()
        c.append(y) 
   
    d = [] #date_firstnonnullline
    for i in c:
        try:
            y = df.loc[i,'Data']
            d.append(y)
        except Exception:
            y = np.nan
            d.append(y)

    e = [] #uniquevalues
    for i in df.columns:
        y = df[i].nunique()
        e.append(y)

    #f= list(map(lambda x,y: (x/y)*100, e,a)) #percentunique

    g = [] #zerocount
    for i in df.columns:
        y = (df[i] == "0").sum()
        g.append(y)
    
    h= list(map(lambda x,y: (x/y)*100, g,a)) #percentzero   
    
    #Selecionando algumas métricas mais relevantes
    estats=pd.DataFrame(data={'percent_missing': b, 'date_firstnonnullline': d, 'percentzero':h },index=df.columns)
    
    return estats.iloc[3:,:] #mostrando as métricas somente para as variáveis de fato

#Mostrando as métricas para toda base de dados
estats(tecnico2)

#Olhando para um ticher em especial
#tecnico2[tecnico2['Ativo']=='ABEV3']

#Olhando para uma data em especial
#tecnico2[tecnico2['Data']=='2001-12-28']

#Mostrando as métricas para um ticker em especial
#estats(tecnico2[tecnico2['Ativo']=='ABEV3'])

#Mostrando as métricas para todos os tickers
#todos= list(map(lambda x: estats(tecnico2[tecnico2['Ativo']==x]), list(tecnico2['Ativo'].unique())))
#estatsall=pd.concat(todos, axis=0, keys=list(tecnico2['Ativo'].unique()))
#estatsall

#Análise visual de missing
#df=tecnico2.copy()
#df = df.sort_values(by=['Data'])

#fontsize = 20
    
#fig, ax = plt.subplots(1, 1, figsize=[10, 150]) #mude aqui o tamanho da matriz
#msno.matrix(df=df.iloc[:,3:], ax=ax, color=(0.2, 0.2, 0.2), sparkline=False, fontsize=fontsize)

#label = list(df['Ativo'].unique())
#ylim_start, ylim_end = ax.get_ylim()
#step_size = df.shape[0] / len(label)
#_ = ax.yaxis.set_ticks(np.arange(ylim_end, ylim_start, step_size))
#_ = ax.yaxis.set_ticklabels(label, fontsize=fontsize)

tecnico2

#Salvando em excel
tecnico2.to_excel('tec.xlsx')
