==> 6_precodiario.ipynb <==
# Dados de Cotações

#Importando pacotes
import pandas as pd
import numpy as np
from urllib.request import urlopen
import missingno as msno
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
from scipy.stats import mstats
warnings.filterwarnings("ignore")

#Descriminando URL de download
url = "https://api.data.economatica.com/1/oficial/datafeed/download/1/oph2LwwiMazuUGjnIuvaibnn%2FLjwrYvs8KW6mxzexjUfc8IXEks1lhwSCNFbVOPf7Kq7k9Ji2icHCj%2BSqTJXZ1V6vxP5eTCjxv2RFETGy6YTlL8gHyw7A7%2BOSCmxQJdlzQYuYjQNL9s6OxAYU48TECrFsZxv%2BSiPZtDHAKiUqi2tS1t5s6V6L1zuWQJ7J1FPvrYivO3JxEZI1kHXWpNKWxy9yQmbPGDbxS6SFMnlmB7Dy86NsSc0V44FV5F0RYZouEbs7FUpVdrxjPHQicvy3WOv8aoS7fnA3M8FDORc2lP5SMPQMNv1BBXWklLoaWiRY93RFFWaY%2BnRR4ouaKgwZQ%3D%3D"
indices = "https://api.data.economatica.com/1/oficial/datafeed/download/1/jIFbIzZ6mj3itlfxKk9y5RW6yi7UT2OuKsIfrip3AbmQn3hIqrNgLKl0NmC3dkb5S4JISEgAD2dNFmrBs7bH%2BSOq4iCd%2FRwDRm9p5x7rtnWyUFAzJJJBQmUMeM48%2BzvYDrkHNHFEkvWGM4ifq%2BJ48jfgvk7dn%2BVGUKkcZxI4FjA4yXM%2FoFjWVpwgeOLBLs%2FpCuttUBMntHZfSWm%2BQv2BR9t%2FwlQPJyvEyvzldFtsC6ugE%2B4EOxGywjw1%2B037t2j0UxZTHoJr1Z4f%2FCXO%2BXGUi1FP7xepIiAmnQLxs3bS8uR9lcTWeRQh9P6EbuqfEY8OC1lz%2B0kAgwoRe65rWMZYHQ%3D%3D"

#%%time
#Abrindo a base
data = pd.read_csv(urlopen(url),encoding='latin-1')
#Retirando sufixo
data['Ativo']=data['Ativo'].str.rstrip('<XBSP>')
#Declarando coluna data como datatime
data['Data'] = pd.to_datetime(data['Data'])
#Renomeando colunas
data.columns = ['ativo','data',"fech","abt","min","max","med","qnegs","qtits","vol"]

#Confirmando o número de empresas
unique=data['ativo'].unique()
#unique
len(unique)

data1=data
#data1

data1=data1[["ativo", "data", "fech"]]
data1[data1 == "-"] = np.nan
data1=data1[data["data"]>"2001-12-30"]
data1=data1.reset_index()
data1=data1.drop(columns=["index"])
#cot=cot.sort_values(['Ativo', "Data"], ascending=[True, True])

data1['ativo'] = data1['ativo'].astype(str)
data1.iloc[:,2:] = data1.iloc[:,2:].astype(float)
data1.dtypes

#REARRANJO PARA O MODELO DE FACTORING
balanco4 = data1
balanco4col = list(range(len(balanco4.iloc[:,2:].columns)))
pv = []
for i in balanco4.iloc[:,2:].columns:
    balanco4col = balanco4[['data','ativo',i]]
    balanco4col = balanco4col.pivot_table(values = i, index = 'data', columns = 'ativo', dropna=False)
    pv.append(balanco4col)

listnames = list(balanco4.iloc[:,2:].columns)
dictionary = dict(zip(listnames, pv))
dictionary.keys()

#dictionary['fech']

#Salvando em excel
pdia=pd.DataFrame(dictionary['fech'])
pdia.reset_index()
pdia

pdia.to_excel('pdia.xlsx')

#%%time
#Abrindo a base
data = pd.read_csv(urlopen(indices),encoding='latin-1')
#Retirando sufixo
data['Ativo']=data['Ativo'].str.rstrip('<XBSP>')
#Declarando coluna data como datatime
data['Data'] = pd.to_datetime(data['Data'])
#Renomeando colunas
data.columns = ['ativo','data',"fech","abt","min","max","med","qnegs","qtits","vol"]

data1=data

data1=data1[["ativo", "data", "fech"]]
data1[data1 == "-"] = np.nan
data1=data1[data["data"]>"2001-12-30"]
data1=data1.reset_index()
data1=data1.drop(columns=["index"])
#cot=cot.sort_values(['Ativo', "Data"], ascending=[True, True])

#data1

data1['ativo'] = data1['ativo'].astype(str)
data1.iloc[:,2:] = data1.iloc[:,2:].astype(float)
data1.dtypes

#REARRANJO PARA O MODELO DE FACTORING
balanco4 = data1
balanco4col = list(range(len(balanco4.iloc[:,2:].columns)))
pv = []
for i in balanco4.iloc[:,2:].columns:
    balanco4col = balanco4[['data','ativo',i]]
    balanco4col = balanco4col.pivot_table(values = i, index = 'data', columns = 'ativo', dropna=False)
    pv.append(balanco4col)

listnames = list(balanco4.iloc[:,2:].columns)
dictionary = dict(zip(listnames, pv))
dictionary.keys()

dictionary['fech']

#Salvando em excel
idia=pd.DataFrame(dictionary['fech'])
idia.reset_index()

idia.columns = ['CDI', 'IBX100', 'IBOV']
idia

idia.to_excel('idia.xlsx')
