==> 7_fatores.ipynb <==
%run ./0_mibrx.ipynb
%run ./1_cotacao.ipynb
%run ./2_balanco.ipynb
%run ./3_financeiro.ipynb
%run ./4_mercado.ipynb
%run ./5_tecnico.ipynb
%run ./6_precodiario.ipynb

#Importando pacotes
import pandas as pd
import numpy as np
from urllib.request import urlopen
import missingno as msno
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
from scipy.stats import mstats
warnings.filterwarnings("ignore")
from  pandas import ExcelWriter


#MATRIZ IBRX
cot=pd.read_excel('cot.xlsx', sheet_name='Sheet1', index_col='data')
cot=cot.drop(columns=["Unnamed: 0"])
cot=cot.reset_index()
cot=cot.sort_values(['ativo', "data"], ascending=[True, True])
bal=pd.read_excel('bal.xlsx', sheet_name='Sheet1', index_col='data')
bal=bal.drop(columns=["Unnamed: 0"])
bal=bal.reset_index()
bal=bal.sort_values(['ativo', "data"], ascending=[True, True])
mer=pd.read_excel('mer.xlsx', sheet_name='Sheet1', index_col='data')
mer=mer.drop(columns=["Unnamed: 0"])
mer=mer.reset_index()
mer=mer.sort_values(['ativo', "data"], ascending=[True, True])
fin=pd.read_excel('fin.xlsx', sheet_name='Sheet1', index_col='data')
fin=fin.drop(columns=["Unnamed: 0"])
fin=fin.reset_index()
fin=fin.sort_values(['ativo', "data"], ascending=[True, True])
tec=pd.read_excel('tec.xlsx', sheet_name='Sheet1', index_col='data')
tec=tec.drop(columns=["Unnamed: 0"])
tec=tec.reset_index()
tec=tec.sort_values(['ativo', "data"], ascending=[True, True])

#Declarando coluna data como datatime
cot['data'] = pd.to_datetime(cot['data'])
bal['data'] = pd.to_datetime(bal['data'])
mer['data'] = pd.to_datetime(mer['data'])
fin['data'] = pd.to_datetime(fin['data'])
tec['data'] = pd.to_datetime(tec['data'])

#conc=pd.concat([cot,bal.iloc[:,2:],mer.iloc[:,2:],fin.iloc[:,2:],tec.iloc[:,2:]], axis=1)

ctc=cot.merge(bal, on=["data", "ativo", "IBRX"]).merge(mer, on=["data", "ativo", "IBRX"]).merge(fin, on=["data", "ativo", "IBRX"]).merge(tec, on=["data", "ativo", "IBRX"])
ctc = ctc.sort_values(by = ['data', "ativo"]).reset_index(drop=True)
ctc

z=len(ctc['ativo'].unique())
z

# #Análise visual de missing
# df=ctc1.copy()
# df = df.sort_values(by=['data'])
# fontsize = 20
# fig, ax = plt.subplots(1, 1, figsize=[6, 6]) #mude aqui o tamanho da matriz
# msno.matrix(df=df.iloc[:,3:], ax=ax, color=(0.2, 0.2, 0.2), sparkline=False, fontsize=fontsize)

datesx = pd.DataFrame(pd.date_range(ctc.data.min(), ctc.data.max(), freq='BM'))
datesx["data"]=datesx[0]
datesx = datesx.drop(columns = 0)
datesy=pd.DataFrame(np.repeat(datesx.values, z, axis=0))
datesy["data"]=datesy[0]
datesy = datesy.drop(columns = 0)
datesy["data"]=pd.to_datetime(datesy["data"])
datesy 

ctc

ctc["data"]=datesy["data"]
ctc

ctc1=ctc[ctc['IBRX']==1]
ctc1

#REARRANJO PARA O MODELO DE FACTORING
balanco4 = ctc1
balanco4col = list(range(len(balanco4.iloc[:,3:].columns)))
pv = []
for i in balanco4.iloc[:,3:].columns:
    balanco4col = balanco4[['data','ativo',i]]
    balanco4col = balanco4col.pivot_table(values = i, index = 'data', columns = 'ativo', dropna=False)
    pv.append(balanco4col)

listnames = list(balanco4.iloc[:,3:].columns)
dictionary = dict(zip(listnames, pv))
dictionary.keys()

writer=pd.ExcelWriter("output.xlsx") 
for key in dictionary:
    dictionary[key].to_excel(writer,sheet_name=key)

writer.save()
writer.close()

valmerc = pd.DataFrame(dictionary['marketcap'])

negoc = pd.DataFrame(dictionary['negocliq6m'])

cols = valmerc.columns.values
bmatrix = pd.DataFrame(index=negoc.index, columns=cols, data=1)

arr = valmerc.to_numpy()
is_equal = ((arr == arr[None].T).any(axis=1))
is_equal[np.tril_indices_from(is_equal)] = False
inds_of_same_cols = [*zip(*np.where(is_equal))]

h = [inds for inds in inds_of_same_cols]
h2 = valmerc.columns[(h)].tolist()

#negoc = dictionary['negc6m']

prematrix = []
for i in h:
    op = pd.DataFrame(negoc.iloc[:,i[0]])
    ap = pd.DataFrame(negoc.iloc[:,i[1]])
    ep = pd.concat([op, ap], axis=1)
    prematrix.append(ep)

#prematrix[0]
    
for v in range(len(prematrix)):
    prematrix[v].insert(0,'new',range(len(prematrix[v])))


for i in range(len(prematrix)):
    for j,a,b in prematrix[i].values:
        #print(j,a,b)
        if a < b:
            bmatrix[prematrix[i].drop(columns = 'new').columns[0]].iloc[int(j)] = 0
            bmatrix[prematrix[i].drop(columns = 'new').columns[1]].iloc[int(j)] = 1
        elif a > b:
            bmatrix[prematrix[i].drop(columns = 'new').columns[0]].iloc[int(j)] = 1
            bmatrix[prematrix[i].drop(columns = 'new').columns[1]].iloc[int(j)] = 0
        elif pd.isna(a):
            bmatrix[prematrix[i].drop(columns = 'new').columns[0]].iloc[int(j)] = 0
        elif pd.isna(b):
            bmatrix[prematrix[i].drop(columns = 'new').columns[1]].iloc[int(j)] = 0
            
            
bmatrix.tail(20)

bmatrix.to_excel('liq.xlsx')

#Compatibilizando datas 
ibrx=pd.read_excel('ibrx.xlsx',sheet_name='Sheet1', index_col='Data')
ibrx1 = ibrx.drop(columns = 'TCSL4')
ibrx1 = ibrx1.reindex(sorted(ibrx1.columns), axis=1)

datest = pd.DataFrame(pd.date_range(ctc.data.min(), ctc.data.max(), freq='BM'))
datest["data"]=datest[0]
datest = datest.drop(columns = 0)
datest

ibrx1.index=datest["data"]
ibrx1.head(13)

ibrx1.to_excel('ibrx1.xlsx')
